Added prices and students for all buildings in buildings.rs and increase the price variable in main.rs 's building struct to i64.

+++ src/main.rs
@@ -16,7 +16,7 @@ struct Building {
     name: &'static str, // The type of building
     students: i32,    // Students per Second that this building generates
     perk_points: i32,// Number of perk points awarded by purchasing this building
-    price: i32, // Price to purchase building
+    price: i64, // Price to purchase building
 }
 
 struct Score {
